apps/candidate_api/pyproject.toml
[tool.poetry]
name = "candidate_api"
version = "0.1.0"
description = "API for managing candidates and materials (Cosmos + Blob)"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
azure-cosmos = "^4.6.0"
azure-storage-blob = "^12.21.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
shared_utils = { path = "../shared_utils", develop = true }
pydantic = "^2.7.0"
pydantic-settings = "^2.2.0"
python-multipart = "^0.0.9"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"

